+++ src/admin_ui/src/admin/widgets/index.ts
@@ -1,0 +1,4 @@
+export * from "./json_field";
+export * from "./json_input";
+export * from "./mapping_field";
+export * from "./mapping_input";