+++ src/admin_ui/src/index.tsx
@@ -1,18 +1,22 @@
 import React from 'react';
-import ReactDOM from 'react-dom';
+import ReactDOM from 'react-dom/client';
 import './index.css';
 import App from './App';
-import * as serviceWorker from './serviceWorker';
-import {AppParams} from "./types";
+import reportWebVitals from './reportWebVitals';
+import {AppParams} from "./admin/types";
 
 declare var __RESTFW_ADMIN_PARAMS__: AppParams;
 
-ReactDOM.render(
-    <App {...__RESTFW_ADMIN_PARAMS__}/>,
-    document.getElementById('root')
+const root = ReactDOM.createRoot(
+    document.getElementById('root') as HTMLElement
+);
+root.render(
+    <React.StrictMode>
+        <App {...__RESTFW_ADMIN_PARAMS__}/>
+    </React.StrictMode>
 );
 
-// If you want your app to work offline and load faster, you can change
-// unregister() to register() below. Note this comes with some pitfalls.
-// Learn more about service workers: https://bit.ly/CRA-PWA
-serviceWorker.unregister();
+// If you want to start measuring performance in your app, pass a function
+// to log results (for example: reportWebVitals(console.log))
+// or send to an analytics endpoint. Learn more: https://bit.ly/CRA-vitals
+reportWebVitals();