+++ src/restfw_admin/restfw_admin/widgets.py
@@ -4,9 +4,9 @@
 :Date: 04.07.2020
 """
 from dataclasses import dataclass, field, fields
-from typing import Any, ClassVar, Dict, List, Literal, Optional, Tuple, Union
+from typing import Any, ClassVar, Dict, List, Literal, Optional, Tuple
 
-from restfw.typing import JsonNumber, SimpleJsonValue
+from restfw.typing import Json, JsonNumber, SimpleJsonValue
 
 from .models import FieldModel
 from .utils import slug_to_title
@@ -84,7 +84,7 @@ class FieldWidget(Widget):
 @dataclass()
 class InputWidget(Widget):
     # Value to be set when the property is null or undefined.
-    initial_value: Optional[SimpleJsonValue] = ra_field('initialValue')
+    initial_value: Optional[Json] = ra_field('defaultValue')
     # Validation rules for the current property.
     validators: Optional[List[Validator]] = None
     # Text to be displayed under the input
@@ -227,7 +227,8 @@ class ChoicesInputWidget(ChoicesWidget, InputWidget):
 @dataclass()
 class SelectInput(ChoicesInputWidget):
     type = 'SelectInput'
-    allow_empty: Optional[bool] = ra_field('allowEmpty')
+    # If the input isn’t required, users can select an empty choice
+    # with an empty_text as label and empty_value as value.
     # Overwrite value of "empty value".
     empty_value: Optional[SimpleJsonValue] = ra_field('emptyValue', default='')
     # The text to display for the empty option.
@@ -360,7 +361,7 @@ class MappingField(Widget):
 
 
 @dataclass()
-class MappingInput(Widget):
+class MappingInput(InputWidget):
     type = 'MappingInput'
     fields: Dict[str, InputWidget] = None
 
@@ -380,44 +381,18 @@ class MappingInput(Widget):
 # Json
 
 @dataclass()
-class JsonOptions:
-    # Name of your root node. Use null or false for no name.
-    name: Optional[str] = None
-    # The indent-width for nested objects, default: 4
-    indent_width: Optional[int] = ra_field('indentWidth')
-    # When set to True, all nodes will be collapsed by default.
-    # Use an integer value to collapse at a particular depth.
-    collapsed: Union[int, bool] = False
-    # When an integer value is assigned, strings will be cut off
-    # at that length. Collapsed strings are followed by an ellipsis.
-    # String content can be expanded and collapsed by clicking on the string value.
-    collapse_strings_after_length: Optional[int] = ra_field('collapseStringsAfterLength')
-    # When set to True, objects and arrays are labeled with size
-    display_object_size: bool = ra_field('displayObjectSize', default=True)
-    # When set to True, data type labels prefix values
-    display_data_types: bool = ra_field('displayDataTypes', default=False)
-
-
-@dataclass()
 class JsonField(FieldWidget):
     type = 'JsonField'
-    add_label: Optional[bool] = ra_field('addLabel', default=True)
-    # Set to True if the value is a string, default: False
-    json_string: Optional[bool] = ra_field('jsonString')
-    # Props passed to react-json-view
-    react_json_options: JsonOptions = ra_field(
-        'reactJsonOptions',
-        default_factory=JsonOptions
-    )
+    # The indent-width for nested objects, default: 2
+    indent_width: Optional[int] = None
 
 
 @dataclass()
-class JsonInput(InputWidget):
+class JsonInput(TextInput):
     type = 'JsonInput'
-    # Set to True if the value is a string, default: False
-    json_string: Optional[bool] = ra_field('jsonString')
-    # Props passed to react-json-view
-    react_json_options: JsonOptions = ra_field(
-        'reactJsonOptions',
-        default_factory=JsonOptions
-    )
+    # The indent-width for nested objects, default: 2
+    indent_width: Optional[int] = None
+    error_text: Optional[str] = None
+    # Set to False if the value is a string in JSON format, default: True
+    parse_json: bool = True
+    multiline: Optional[bool] = True